+++ Demo/Demo/SwiftUIComponents/SwiftUIPaymentButtonDemo.swift
@@ -1,5 +1,5 @@
 import SwiftUI
-import PayPalUI
+import PaymentButtons
 
 struct SwiftUIPaymentButtonDemo: View {
 
@@ -8,18 +8,22 @@ struct SwiftUIPaymentButtonDemo: View {
 
     @State private var fundingIndex = 0
     private var fundingSources = PaymentButtonFundingSource.allCasesAsString()
-    @State private var selectedFunding = PaymentButtonFundingSource.allCases()[0]
+    @State private var selectedFunding = PaymentButtonFundingSource.allCases[0]
 
     @State private var colorsIndex = 0
     @State private var colors = PayPalButton.Color.allCasesAsString()
 
     @State private var edgesIndex = 0
     private var edges = PaymentButtonEdges.allCasesAsString()
-    @State private var selectedEdge = PaymentButtonEdges.allCases()[0]
+    @State private var selectedEdge = PaymentButtonEdges.allCases[0]
 
     @State private var sizesIndex = 1
     private var sizes = PaymentButtonSize.allCasesAsString()
-    @State private var selectedSize = PaymentButtonSize.allCases()[1]
+    @State private var selectedSize = PaymentButtonSize.allCases[1]
+    
+    @State private var labelIndex = 0
+    private var labels = PayPalButton.Label.allCasesAsString()
+    @State private var selectedLabel = PayPalButton.Label.allCases[0]
 
     @ViewBuilder
     var body: some View {
@@ -32,7 +36,7 @@ struct SwiftUIPaymentButtonDemo: View {
                 }
                 .pickerStyle(SegmentedPickerStyle())
                 .onChange(of: fundingIndex) { _ in
-                    selectedFunding = PaymentButtonFundingSource.allCases()[fundingIndex]
+                    selectedFunding = PaymentButtonFundingSource.allCases[fundingIndex]
                     colors = getColorFunding(with: selectedFunding)
                     colorsIndex = 0
                     pickerId += 1 // Workaround to change ID of picker. ID is updated to force refresh, https://developer.apple.com/forums/thread/127560
@@ -57,7 +61,7 @@ struct SwiftUIPaymentButtonDemo: View {
                 }
                 .pickerStyle(SegmentedPickerStyle())
                 .onChange(of: edgesIndex) { _ in
-                    selectedEdge = PaymentButtonEdges.allCases()[edgesIndex]
+                    selectedEdge = PaymentButtonEdges.allCases[edgesIndex]
                     buttonId += 1
                 }
 
@@ -68,22 +72,35 @@ struct SwiftUIPaymentButtonDemo: View {
                 }
                 .pickerStyle(SegmentedPickerStyle())
                 .onChange(of: sizesIndex) { _ in
-                    selectedSize = PaymentButtonSize.allCases()[sizesIndex]
+                    selectedSize = PaymentButtonSize.allCases[sizesIndex]
                     buttonId += 1
                 }
 
                 switch selectedFunding {
                 case .payPal:
+                    if selectedSize == .expanded || selectedSize == .full {
+                        Picker("label", selection: $labelIndex) {
+                            ForEach(labels.indices, id: \.self) { index in
+                                Text(labels[index])
+                            }
+                        }
+                        .pickerStyle(SegmentedPickerStyle())
+                        .onChange(of: labelIndex) { _ in
+                            selectedLabel = PayPalButton.Label.allCases[labelIndex]
+                            buttonId += 1
+                        }
+                    }
                     PayPalButton.Representable(
-                        color: PayPalButton.Color.allCases()[colorsIndex],
+                        color: PayPalButton.Color.allCases[colorsIndex],
                         edges: selectedEdge,
-                        size: selectedSize
+                        size: selectedSize,
+                        label: selectedLabel
                     )
                     .id(buttonId)
 
                 case .payLater:
                     PayPalPayLaterButton.Representable(
-                        color: PayPalPayLaterButton.Color.allCases()[colorsIndex],
+                        color: PayPalPayLaterButton.Color.allCases[colorsIndex],
                         edges: selectedEdge,
                         size: selectedSize
                     )
@@ -91,7 +108,7 @@ struct SwiftUIPaymentButtonDemo: View {
 
                 case .credit:
                     PayPalCreditButton.Representable(
-                        color: PayPalCreditButton.Color.allCases()[colorsIndex],
+                        color: PayPalCreditButton.Color.allCases[colorsIndex],
                         edges: selectedEdge,
                         size: selectedSize
                     )
@@ -112,5 +129,12 @@ struct SwiftUIPaymentButtonDemo: View {
         case .credit:
             return PayPalCreditButton.Color.allCasesAsString()
         }
+    }
+}
+
+struct SwiftUIPaymentButtonDemo_Preview: PreviewProvider {
+    
+    static var previews: some View {
+        SwiftUIPaymentButtonDemo()
     }
 }