+++ UnitTests/FraudProtectionTests/MockMagnesSDKResult.swift
@@ -1,0 +1,15 @@
+import Foundation
+@testable import FraudProtection
+
+class MockMagnesSDKResult: MagnesSDKResult {
+
+    var payPalClientMetaDataId: String
+
+    init(payPalClientMetaDataId: String = "") {
+        self.payPalClientMetaDataId = payPalClientMetaDataId
+    }
+
+    func getPayPalClientMetaDataId() -> String {
+        return payPalClientMetaDataId
+    }
+}